Reorganize makeFunctions.ts and mocks.ts into helpers/ dir. Add helpers/ into jest ignore
