+++ bloggies_backend/src/__tests__/helpers/mocks.ts
@@ -1,29 +1,29 @@
 export const TEST_EMAIL = "test@test.com";
 export const TEST_SECONDARY_EMAIL = "secondarytest@test.com";
 export const TEST_DISPLAY_NAME = "milkteafan";
-export const TEST_SECONDARY_DISPLAY_NAME = "mochafan"
+export const TEST_SECONDARY_DISPLAY_NAME = "mochafan";
 export const MOCK_USER_ID = 1;
 export const VALID_CARD = {
   number: '4242424242424242',
   exp_month: 3,
   exp_year: 2022,
   cvc: '314',
-}
+};
 export const INVALID_CARD = {
   number: '4000000000000002',
   exp_month: 3,
   exp_year: 2025,
   cvc: '314'
-}
+};
 export const VALID_FAIL_CHARGE_CARD = {
   number: '4000000000000341',
   exp_month: 3,
   exp_year: 2025,
   cvc: '314'
-}
+};
 export const INVALID_CUSTOMER_ID = "cus_notarealcustomerid";
 export const INVALID_SUB_ID = "sub_notarealsubid";
 export const TEST_POST_TITLE = "Strawberry Basil Soda";
 export const TEST_POST_DESC = "I like strawberry basil soda";
 export const TEST_POST_BODY = "Strawberry basil soda is cool xd uwu <3";
-export const TEST_COMMENT_BODY = "This is a really great post about strawberry soda"
+export const TEST_COMMENT_BODY = "This is a really great post about strawberry soda";
